+++ src/findMailbox.ts
@@ -1,0 +1,11 @@
+import { Db, Datum } from './database';
+
+export const findMailbox = (database: Db, repoId: number): Datum | null => {
+  const mailbox = database.data.find(({ repositories }) => repositories.some(id => id === repoId));
+
+  if (mailbox === undefined) {
+    return null;
+  }
+
+  return mailbox;
+};
